using vpn in tmux again?

+++ zsh/zsh/2-tmux.sh
@@ -4,24 +4,12 @@ function set_tmux_window_name() {
 
 function work() {
     [[ -n $TMUX ]] && OP="switch" || OP="attach"
-    echo $OP
-    if ! $(tmux ls | grep -q work); then
-        echo "new session!"
-        tmux new-session -d -s work
-#         tmux rename-window -t $session:1 "todo"
-#         tmux send-keys -t "todo" "todo" C-m
-#         tmux new-window -t $session:2 -n "puppet"
-#         tmux send-keys -t "puppet" "vip" C-m
-#         tmux new-window -t $session:3 -n "zsh"
-#         tmux attach -t "$session:zsh"
-    fi
+    $(tmux ls | grep -q work) || tmux new-session -d -s work
     tmux $OP -t work
 }
 
 function unwork() {
-    if $(tmux ls | grep -q work); then
-        tmux kill-session -t work
-    fi
+    $(tmux ls | grep -q work) && tmux kill-session -t work
 }
 
 function tmux_lcrown() {
